# back/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API - Agenciamento de Transferência de Jogadores",
    "description": "Documentação da API do sistema de agenciamento de transferência de jogadores para clubes. Rotas protegidas exigem um token JWT: faça login em /clubes/login ou /admin/login, copie o token e clique em \"Authorize\" abaixo, informando: Bearer SEU_TOKEN",
    "version": "1.0.0"
  },
  "basePath": "/",
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Informe: Bearer SEU_TOKEN (token retornado por /clubes/login ou /admin/login)"
    }
  },
  "paths": {
...
    "/clubes/": {
      "post": {
        "tags": [
          "Clubes"
        ],
        "summary": "Cadastra um clube",
        "description": "Cria um novo clube (cliente do sistema), que poderá fazer login e enviar propostas por jogadores.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "FC Exemplo"
                },
                "pais": {
                  "type": "string",
                  "example": "Brasil"
                },
                "email": {
                  "type": "string",
                  "example": "diretoria@fcexemplo.com"
                },
                "senha": {
                  "type": "string",
                  "example": "123456"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Clube cadastrado com sucesso."
          },
          "400": {
            "description": "Dados inválidos ou e-mail já utilizado."
          }
        }
      }
    },
    "/clubes/login": {
      "post": {
        "tags": [
          "Clubes"
        ],
        "summary": "Login do clube",
        "description": "Autentica o clube e retorna um token JWT. O front-end guarda o token (em memória e, se \"Manter Conectado\", também no LocalStorage) e o envia no cabeçalho Authorization das próximas requisições.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "diretoria@fcexemplo.com"
                },
                "senha": {
                  "type": "string",
                  "example": "123456"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login realizado com sucesso."
          },
          "400": {
            "description": "E-mail ou senha inválidos."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/clubes/me": {
      "get": {
        "tags": [
          "Clubes"
        ],
        "summary": "Dados do clube autenticado",
        "description": "Retorna os dados do clube dono do token enviado no cabeçalho Authorization. Usada para restaurar a sessão ao reabrir o front-end.",
        "responses": {
          "200": {
            "description": "Clube encontrado."
          },
          "401": {
            "description": "Não autenticado."
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/admin/login": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Login do administrador",
        "description": "Autentica o administrador e retorna um token JWT, usado no cabeçalho Authorization para acessar a área restrita.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "admin@agencia.com"
                },
                "senha": {
                  "type": "string",
                  "example": "admin123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login realizado com sucesso."
          },
          "400": {
            "description": "E-mail ou senha inválidos."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/dashboard": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Dados da Dashboard",
        "description": "Retorna as contagens gerais e as agregações usadas nos gráficos da área restrita. Requer autenticação de administrador.",
        "responses": {
          "200": {
            "description": "Dados da Dashboard retornados com sucesso."
          },
          "401": {
            "description": "Não autenticado."
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  }
}